fixCustomindor

+++ DATN/src/main/java/webbangiaydabong/repository/CustommerInfoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import webbangiaydabong.entity.CustommerInfo;
 
-public interface CustommerInfoRepository extends JpaRepository<CustommerInfo, Integer>{
+public interface CustommerInfoRepository extends JpaRepository<CustommerInfo, Long>{
 
 }
